requirements.txt
requests
selenium
python-dotenv
cryptography
google-generativeai
